- refactored to reunite all handlers in the same files - removed unused imports here and there - cleaned up requirements.txt
