+++ defrag/__main__.py
@@ -14,6 +14,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 
+from defrag.routes import expose_modules_to_handlers
 from defrag.modules.helpers.requests import Req
 from defrag.modules.db.redis import RedisPool
 import uvicorn
@@ -42,12 +43,16 @@ def main() -> None:
 
 @app.on_event("startup")
 async def register_modules_as_services() -> None:
-    """ Registers all modules implementing 'register_service()'. """
+
+    # Registers all modules implementing 'register_service()'.
     with RedisPool() as conn:
         conn.flushall()
     for service in IMPORTED.values():
         if hasattr(service, "register_service"):
             service.register_service()
+
+    # Propagate modules names to route handlers
+    expose_modules_to_handlers(list(IMPORTED.keys()))
 
 
 @app.on_event("shutdown")
