+++ defrag/modules/bugs.py
@@ -1,12 +1,10 @@
 import asyncio
-from functools import partial
 from bugzilla.base import Bugzilla
 from pydantic.main import BaseModel
-from defrag.modules.helpers import Query, CacheQuery, QueryResponse
-from defrag.modules.helpers.services_manager import Run, ServiceTemplate, ServicesManager
+from defrag.modules.helpers.services_manager import ServiceTemplate, ServicesManager
 from defrag.modules.helpers.cache_stores import CacheStrategy, DStore, RedisCacheStrategy
 from typing import Any, Dict, List, Optional, Union
-from defrag import app, BUGZILLA_USER, BUGZILLA_PASSWORD
+from defrag import BUGZILLA_USER, BUGZILLA_PASSWORD
 from defrag.modules.helpers.sync_utils import as_async
 from defrag.modules.helpers.exceptions import ParsingException
 from defrag.modules.helpers.requests import Req
@@ -163,33 +161,3 @@ def register_service():
     # sends everything to the ServicesManager for registration
     ServicesManager.register_service(__MOD_NAME__, service)
 
-
-@app.get("/" + __MOD_NAME__ + "/bug/{bug_id}")
-async def get_bug(bug_id: int) -> QueryResponse:
-    # declares how this request should interface with the cache
-    cache_query = CacheQuery(service="bugs", item_key=bug_id)
-    # declares what function to run if the item the request is looking for
-    # cannot find it in the cache store
-    fallback = partial(get_this_bug, bug_id)
-    # run the request
-    return await Run.query(cache_query, fallback)
-
-
-@app.get("/" + __MOD_NAME__ + "/")
-async def root() -> QueryResponse:
-    return QueryResponse(query="info", results=[
-                         {"module": "Bugzilla", "description": "Get information about bugs on bugzilla.opensuse.org"}])
-
-
-@app.get("/" + __MOD_NAME__ + "/search/")
-async def search(term: str) -> QueryResponse:
-    query = BugzillaQueryEntry(search_string=term)
-    result = await search_all_bugs(query)
-    # This is not as fancy as it was before, but now it actually works.
-    # Plus, before id didn't cache anyway, so this should be fine. We can
-    # still make it better in the future
-    return QueryResponse(
-        query=Query(
-            service="bugs"),
-        results_count=len(result),
-        results=result)
