+++ defrag/modules/docs.py
@@ -1,7 +1,5 @@
-from defrag import app
 from defrag.modules.helpers.services_manager import ServiceTemplate, ServicesManager
 from defrag.modules.helpers.requests import Req
-from defrag.modules.helpers import Query, QueryResponse
 from lunr import lunr
 from lunr.index import Index
 from bs4 import BeautifulSoup
@@ -128,29 +126,6 @@ def search_indexes_in_parallel(keywords: str) -> List[Dict[str, Any]]:
         tw_worker = executor.submit(
             search_index, **{"idx": indexes["tumbleweed"]["index"], "source": "tumbleweed", "keywords": keywords})
         return leap_worker.result() + tw_worker.result()
-
-
-@app.get("/" + __MOD_NAME__ + "/single/{source}/")
-async def search_single_source_docs(source: str, keywords: str) -> QueryResponse:
-    if not ready_to_index([source]):
-        if source == "tumbleweed":
-            set_global_index("tumbleweed", make_tumbleweed_index(await get_data(source)))
-        else:
-            set_global_index("leap", make_leap_index(await get_data(source)))
-    results = sorted_on_score(search_index(
-        indexes[source]["index"], source, keywords))
-    return QueryResponse(query=Query(service="search_docs"), results_count=len(results), results=results)
-
-
-@app.get("/" + __MOD_NAME__ + "/merged/")
-async def search(keywords: str) -> QueryResponse:
-    if not ready_to_index(["leap", "tumbleweed"]):
-        results = await make_search_set_indexes_in_parallel(keywords)
-        return QueryResponse(query=Query(service="search_docs"), results_count=len(results), results=results)
-    else:
-        results = sorted_on_score(search_indexes_in_parallel(keywords))
-        return QueryResponse(query=Query(service="search_docs"), results_count=len(results), results=results)
-
 
 def register_service():
     asyncio.create_task(make_search_set_indexes_in_parallel(""))
