+++ defrag/modules/organizer.py
@@ -1,11 +1,10 @@
 import asyncio
 from defrag.modules.helpers.data_manipulation import partition_left_right
 from functools import reduce
-from defrag import app
 from datetime import datetime, timedelta, timezone
 from defrag.modules.helpers.sync_utils import as_async
 from defrag.modules.db.redis import RedisPool
-from defrag.modules.helpers import EitherErrorOrOk, FailuresAndSuccesses, Query, QueryResponse
+from defrag.modules.helpers import EitherErrorOrOk, FailuresAndSuccesses
 from defrag.modules.dispatcher import Dispatcher, Dispatchable, Notification, TelegramNotification
 from pottery import RedisDict
 from typing import Any, Dict, List, Optional, Tuple
@@ -354,67 +353,3 @@ class Calendar:
         def rendering() -> List[Dict[str, Any]]:
             return [e for e in Calendar.container.values() if condition(e)]
         return await as_async(rendering)()
-
-# --------
-# HANDLERS
-# --------
-
-
-@app.post(f"/{__MOD_NAME__}/add_reminder/")
-async def post_reminders(reminder: Reminders) -> QueryResponse:
-    query = Query(service=__MOD_NAME__)
-    if not reminder.tgt:
-        return QueryResponse(query=query, error=f"You need to add a 'tgt' (datetime string encoded as {FORMAT} to set this reminder: {reminder}")
-    await Reminders.schedule(tgt=reminder.tgt, notification=reminder.notification, deltas=reminder.deltas)
-    return QueryResponse(query=Query(service=__MOD_NAME__), message="Reminder(s) set!")
-
-
-@app.post(f"/{__MOD_NAME__}/add_reminder_for/")
-async def post_reminders_for(event_id: int, reminders: Reminders) -> QueryResponse:
-    query = Query(service=__MOD_NAME__)
-    reply = "Calendar is empty!"
-    if not Calendar.container:
-        return QueryResponse(query=query, error=reply)
-
-    def look_up(event_id): 
-        return Calendar.container[event_id] if event_id in Calendar.container.keys() else None
-    found = await as_async(look_up)(event_id)
-    if not found:
-        reply = f"Unable to find this calendar item: {event_id}"
-        return QueryResponse(query=query, error=reply)
-    await Reminders.schedule(tgt=found["start"], notification=reminders.notification, deltas=reminders.deltas)
-    return QueryResponse(query=query, message=f"Thanks, reminders set for {event_id}")
-
-
-@app.post(f"/{__MOD_NAME__}/add_fedocal_events/")
-async def post_fedocal_events(events: List[FedocalEvent], reminders: Reminders) -> QueryResponse:
-    query = Query(service=__MOD_NAME__)
-    results = await Calendar.add_all_new_events(events=[event_from_fedocal(m) for m in events], notification=reminders.notification, deltas=reminders.deltas)
-    keys = [e.ok["id"] for e in results.successes if hasattr(e, "ok")]
-    return QueryResponse(query=query, message="event(s) added and reminder(s) set!", results=keys, results_count=len(keys))
-
-
-@app.post(f"/{__MOD_NAME__}/add_events/")
-async def post_events(events: List[CustomEvent], reminders: Reminders) -> QueryResponse:
-    query = Query(service=__MOD_NAME__)
-    results = await Calendar.add_all_new_events(events=events, notification=reminders.notification, deltas=reminders.deltas)
-    keys = [e.ok["id"] for e in results.successes if hasattr(e, "ok")]
-    res = QueryResponse(query=query, message="event(s) added and reminder(s) set!", results=keys, results_count=len(keys))
-    return res
-
-
-@app.post(f"/{__MOD_NAME__}/cancel_event/")
-async def post_cancel_event(event_id: str) -> QueryResponse:
-    query = Query(service=__MOD_NAME__)
-    result = await Calendar.cancel(event_id)
-    if hasattr(result, "ok"):
-        return QueryResponse(query=query, message=f"event and reminder(s) cancelled for {event_id}")
-    else:
-        return QueryResponse(query=query, message=f"Unable to cancel {event_id}")
-
-
-@app.get(f"/{__MOD_NAME__}/calendar/")
-async def get_calendar(start: str, end: str) -> QueryResponse:
-    query = Query(service=__MOD_NAME__)
-    results = await Calendar.render(start_str=start, end_str=end)
-    return QueryResponse(query=query, results=results, results_count=len(results))
