+++ defrag/modules/wikis.py
@@ -1,7 +1,5 @@
 from pydantic.main import BaseModel
-from defrag.modules.helpers import Query, QueryResponse
 from defrag.modules.helpers.requests import Req
-from defrag import app
 from typing import List
 from dateutil import parser
 
@@ -41,8 +39,3 @@ async def search_wikis_as_gen(keywords: str) -> List[WikiGenEntry]:
         return []
 
 
-@app.get(f"/{__MOD_NAME__}/search/")
-async def search(keywords: str) -> QueryResponse:
-    results = await search_wikis_as_list(keywords)
-    query = Query(service=__MOD_NAME__)
-    return QueryResponse(query=query, results=results, results_count=len(results))
