+++ requirements.txt
@@ -6,25 +6,32 @@
 #
 
 -i https://pypi.org/simple
-asgiref==3.4.1
-click==8.0.1
-fastapi==0.66.0
-h11==0.12.0
-httptools==0.2.0
-mmh3==3.0.0
-pottery==1.3.1
-pydantic==1.8.2
-python-dotenv==0.18.0
-python-twitter==3.5
-PyYAML==5.4.1
-redis==3.5.3
-starlette==0.14.2
-typing-extensions==3.10.0.0
+asgiref==3.4.1; python_version >= '3.6'
+atoma==0.0.17
+attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
+beautifulsoup4==4.10.0
+certifi==2021.5.30
+charset-normalizer==2.0.6; python_version >= '3'
+click==8.0.1; python_version >= '3.6'
+defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
+fastapi==0.68.1
+future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
+h11==0.12.0; python_version >= '3.6'
+idna==3.2; python_version >= '3'
 lunr==0.6.0
-urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
-yarl==1.6.3; python_version >= '3.6'
+mmh3==3.0.0
+oauthlib==3.1.1; python_version >= '3.6'
+pottery==1.4.1
+pydantic==1.8.2; python_full_version >= '3.6.1'
+python-bugzilla==3.1.0
+python-dateutil==2.8.2
+python-twitter==3.5
+redis==3.5.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
+requests-oauthlib==1.3.0
+requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
+six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
+soupsieve==2.2.1; python_version >= '3.6'
+starlette==0.14.2; python_version >= '3.6'
+typing-extensions==3.10.0.2
+urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
 uvicorn==0.15.0
-Pyrogram==1.2.9
-async-lru==1.0.2
-TgCrypto==1.2.2
-python-decouple==3.4
